pyproject.toml
[tool.poetry]
name = "smartthings-rest"
version = "0.1.3"
description = "Smart and straightforward lib for controlling things with smartthings"
authors = ["Viktor Freiman <freiman.viktor@gmail.com>"]
license = "MIT"
repository = "https://github.com/viktorfreiman/smartthings-rest"
documentation = "https://smartthings-rest.readthedocs.io"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.6.2"
requests = "^2.26.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
flake8-bugbear = "^21.9.2"
flake8-quotes = "^3.3.1"
Pillow = "^8.4.0"
rinohtype = "^0.5.3"
Sphinx = "^4.2.0"
sphinx-argparse = "^0.3.1"
sphinx-autobuild = "^2021.3.14"
sphinx-rtd-theme = "^1.0.0"
sphinxcontrib-kroki = "^1.3.0"
sphobjinv = "^2.1"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
